Chore: move navigation components into `components/` directory - Rename `nav/` directory to `Navigation/`. - Use absolute imports for assets in `DesktopNav.jsx` and `MobileNav.jsx`. - Update import for navigation elements in `App.jsx`.
